long 2

+++ src/main/java/com/example/demo/api/v1/BannerRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 
-public interface BannerRepository extends JpaRepository<Banner, Integer> {
-  Optional<Banner> findById(Integer id);
+public interface BannerRepository extends JpaRepository<Banner, Long> {
+  Optional<Banner> findById(Long id);
     //Banner findByName(String name);
 }
